server/requirements.txt
flask
flask-cors
langchain
langchain-google-genai
google-generativeai
scikit-learn
